+++ attributes/default.rb
@@ -1,2 +1,2 @@
 # Listen address for the privoxy service
-default['privoxy-nativex']['listen-address'] = node['ipaddress']
+default['privoxy-simplyadrian']['listen-address'] = node['ipaddress']