+++ recipes/default.rb
@@ -1,8 +1,8 @@
 #
-# Cookbook Name:: privoxy-nativex
+# Cookbook Name:: privoxy-simplyadrian
 # Recipe:: default
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
@@ -11,7 +11,7 @@ include_recipe	"privoxy"
 
 begin
   r = resources(:template => "/etc/privoxy/config")
-  r.cookbook "privoxy-nativex"
+  r.cookbook "privoxy-simplyadrian"
 rescue Chef::Exceptions::ResourceNotFound
   Chef::Log.warn "could not find template to override!"
 end