+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-privoxy-nativex CHANGELOG
+privoxy-simplyadrian CHANGELOG
 =========================
 
-This file is used to list changes made in each version of the privoxy-nativex cookbook.
+This file is used to list changes made in each version of the privoxy-simplyadrian cookbook.
 
 0.1.0
 -----
-- Adrian Herrera - Initial release of privoxy-nativex
+- Adrian Herrera - Initial release of privoxy-simplyadrian
 
 - - -